# util/cmpver: faster Compare (-85% runtime)